\name{getDir}
\alias{getDir}
\title{
	Attain the name of a directory
}
\description{
	An internal function that creates a dialogue box to attain the name of a directory to be created.
}
\usage{
getDir(window)
}
\arguments{
  \item{window}{
	The window the dialogue box resides in.  
}
}

\value{
 \item{Dir}{The name of a directory.}
}

\author{
Joe Swintek
}
