\name{openCB}
\alias{openCB}
\title{
	Open a csv file
}
\description{
	An internal function that creates a dialogue box used to open a file with a csv extension.   
}
\usage{
openCB(window)
}
\arguments{
  \item{window}{
	The window the dialogue box resides in. 
}
}
\value{
 \item{df}{The data frame that is opened.}
}
\author{
	Joe Swintek
}