\name{plotPath}
\alias{plotPath}

\title{
	Plot histopathological data
}
\description{
	GUI for plotting histopathological data.
}
\usage{
plotPath()
}

\details{
	See \code{\link{plotRSCABS}} or \code{\link{setPlotParms}} for further details. 
}

\author{
	Joe Swintek
}

\seealso{
\code{\link{plotRSCABS}}
}

\keyword{Plotting}

