\name{runEffectSelect}
\alias{runEffectSelect}
\title{
	Runs a detailed analysis on an endpoint
}
\description{
	An internal function called when the "Get Details on a Response" button is pushed.  
	This function sanitizes the user's input, performs a detailed analysis on an endpoint, then prints them to the display.    
}
\usage{
runEffectSelect()
}


\author{
	Joe Swintek
}

