\name{saveCB}
\alias{saveCB}
\title{
	Attain the name of a file to be saved
}
\description{
	An internal function used to call a dialogue box to name a file to be saved.
}
\usage{
saveCB(window, OutData)
}

\arguments{
  \item{window}{
	The window the dialogue box resides in.  
}
  \item{OutData}{
	Data to be saved.
}
}


\author{
	Joe Swintek
}

