\name{saveFile}
\alias{saveFile}
\title{
	Saves a data set
}
\description{
	An internal function that save a data set as csv file.  
}

\usage{
saveFile(File, OutData)
}

\arguments{
  \item{File}{
	Name of the file to be saved.
}
  \item{OutData}{
	Data to be saved.
}
}

\author{
	Joe Swintek
}

